+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/combat/fletching/entity/projectile/QuartzArrow.java
@@ -1,7 +1,12 @@
 package insane96mcp.iguanatweaksexpanded.module.combat.fletching.entity.projectile;
 
 import insane96mcp.iguanatweaksexpanded.module.combat.fletching.Fletching;
+import insane96mcp.iguanatweaksreborn.module.combat.PiercingDamage;
+import insane96mcp.iguanatweaksreborn.utils.MCUtils;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
@@ -9,12 +14,33 @@ import net.minecraft.world.level.Level;
 public class QuartzArrow extends ITEArrow {
     public QuartzArrow(EntityType<? extends Arrow> pEntityType, Level pLevel) {
         super(pEntityType, pLevel);
-        this.friction = 0.995f;
-        this.velocityMultiplier = 1.25f;
     }
 
     @Override
     protected ItemStack getPickupItem() {
         return new ItemStack(Fletching.QUARTZ_ARROW_ITEM.get());
+    }
+
+    @Override
+    protected void doPostHurtEffects(LivingEntity entity) {
+        super.doPostHurtEffects(entity);
+        Entity entity1 = this.getOwner();
+        DamageSource damageSource;
+        if (entity1 == null) {
+            damageSource = this.damageSources().source(PiercingDamage.PIERCING_MOB_ATTACK, this);
+        } else {
+            damageSource = this.damageSources().source(PiercingDamage.PIERCING_MOB_ATTACK, entity1);
+            if (entity1 instanceof LivingEntity) {
+                ((LivingEntity)entity1).setLastHurtMob(entity);
+            }
+        }
+        float damage = (float) ((0.1f + this.getBaseDamage() * 0.1f) * this.getDeltaMovement().length());
+
+        MCUtils.attackEntityIgnoreInvFrames(damageSource, damage, entity, null, true);
+    }
+
+    @Override
+    public double getBaseDamage() {
+        return super.getBaseDamage();
     }
 }