+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/world/seasons/Seasons.java
@@ -78,11 +78,11 @@ public class Seasons extends Feature {
 
 	@Config(min = 0)
 	@Label(name = "Time Control day night duration", description = "How many minutes will day and night duration be (with this set to 10 the day will last 10 minutes and the night 10 minutes for a grand total of 20 minutes). This also adjusts the day_duration config option in Serene Seasons. Vanilla is 10.")
-	public static Double timeControlDayNightDuration = 10d;
+	public static Double timeControlDayNightDuration = 15d;
 
 	@Config(min = 0)
 	@Label(name = "Time Control day night shift", description = "How many minutes will day and night duration be shifted based off seasons? E.g. in Mid spring / autumn the duration of day and night is vanilla, when moving off those seasons day and night will last this many minutes more/less. In mid summer / winter the duration of day and night duration will be more / less by 3 times this value. Set to 0 to disable")
-	public static Double timeControlDayNightShift = 1d;
+	public static Double timeControlDayNightShift = 2d;
 
 	/*@Config
 	@Label(name = "Remove SS TimeSkipHandler", description = "There's a strange bug with the TimeSkipHandler in Serene Seasons where on server the season changed event is still triggered even if no player's online and with random seasons. If this is enabled, the progress_season_while_offline config option is set to false and there's no player online, the TimeSkipHandler will not be triggered.")