+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Quartz arrows now deal bonus piercing damage instead of flying faster and farther
+
 ## 1.12.2
 * Updated advancements
   * Nether, End and Farmers' Delight advancements are now merged into the main advancements of the mod (no longer have their own tab)