+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 * Quartz arrows now deal bonus piercing damage instead of flying faster and farther
+* Fixed part breaker dropping multiple items that already dropped naturally
+  * E.g. Wither Skeletons can no longer drop 2 skulls
 
 ## 1.12.2
 * Updated advancements