+++ changelog.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## Upcoming
-* Quartz arrows now deal bonus piercing damage instead of flying faster and farther
+* Quartz arrows 
+  * Now deal bonus piercing damage 
+  * Base damage: ~~2~~ -> 1
+  * No longer fly faster and farther
 * Fixed part breaker dropping multiple items that already dropped naturally
   * E.g. Wither Skeletons can no longer drop 2 skulls
 * Serene Seasons' seasonal_crops is no longer set to false, as it's done in ITR