+++ changelog.md
@@ -9,6 +9,7 @@
   * E.g. Wither Skeletons can no longer drop 2 skulls
 * Decreased Gravity Defying cost (~~5~~ -> 4)
 * Serene Seasons' seasonal_crops is no longer set to false, as it's done in ITR
+* Serene Seasons' ice melting fix has been moved to ITR
 
 ## 1.12.2
 * Updated advancements