+++ changelog.md
@@ -8,6 +8,7 @@
 * Fixed part breaker dropping multiple items that already dropped naturally
   * E.g. Wither Skeletons can no longer drop 2 skulls
 * Decreased Gravity Defying cost (~~5~~ -> 4)
+* Increased day/night cycle duration (vanilla is 20 minutes, has been increased to 30)
 * Serene Seasons' seasonal_crops is no longer set to false, as it's done in ITR
 * Serene Seasons' ice melting fix has been moved to ITR